Disable TripleBuffer to reduce YouTube frame drops
TripleBuffer is enabled by default, but seemed to be causing upward of
50% frame drops (along with high CPU usage) when playing YouTube from
within a browser. mpv didn't seem to exhibit the same symptoms. Not sure
if this issue coincided with upgrading from Debian bookworm to trixie,
but setting `Option "TripleBuffer" "false" seems to have done the trick
for now.

Have also compared with my desktop's performance with nvidia-driver,
which continues to track well.